change in polynomial_inerval_fit: integer indices (i.e. equally-spaced measurements) -> double values
tracking_goertzel_filter: reset more easily after the filter did not lock on something

+++ trunk/HFMonitor/src/sdr/perseus/fx2_control.cpp
@@ -203,7 +203,7 @@ namespace Perseus {
   protected:
     virtual eeprom_data read_eeprom_lowlevel(boost::uint16_t addr, size_t length) {
       int transferred(0);
-      const EEPROM::cmd cmd = {CMD::eeprom_read, addr, length};
+      const EEPROM::cmd cmd = {CMD::eeprom_read, addr, uint8_t(length)};
       ASSERT_THROW(_usb_control->submit_bulk(EndPoint::cmd,
                                              (unsigned char *)&cmd, sizeof(cmd),
                                              &transferred, FX2::usb_timeout) == 0);
